+++ hybrid-analysis/operations.py
@@ -28,11 +28,12 @@ GET_ENVIRONMENTS = '/api/v2/system/environments'
 GET_QUOTA = '/api/v2/key/submission-quota'
 KEY_LIMITS = '/api/v2/key/current'
 SEARCH_HASHES = '/api/v2/search/hash'
-SAMPLE_DROPPED_FILES = '/api/v2/report/{ID}/dropped-files'
+SAMPLE_DROPPED_FILES = '/api/v2/report/{ID}/dropped-files-v2'
 SAMPLE_SCREENSHOTS = '/api/v2/report/{ID}/screenshots'
 SEARCH = '/api/v2/search/terms'
 SUBMISSION_STATE = '/api/v2/report/{ID}/state'
 URL_QUICK_SCAN = '/api/v2/quick-scan/url'
+QUICK_SCAN_BY_ID = '/api/v2/quick-scan/{ID}'
 SUBMIT_URL = '/api/v2/submit/url'
 
 ACTION_SCRIPT = {
@@ -358,6 +359,17 @@ def url_quick_scan(config, params):
         raise ConnectorError(Err)
 
 
+def quick_scan_by_id(config, params):
+    try:
+        ID = params.get('id')
+        QUICK_SCAN_ID = QUICK_SCAN_BY_ID.format(ID=ID)
+        response = _api_request("get", QUICK_SCAN_ID, config)
+        return response
+    except Exception as Err:
+        logger.exception("Fail : {}".format(str(Err)))
+        raise ConnectorError(str(Err))
+
+
 def hashes_search(config, params):
     try:
         hash_value = params.get('hashCode')
@@ -499,6 +511,7 @@ hybrid_analysis_ops = {
     'get_feed': get_feed,
     'hashes_search': hashes_search,
     'url_quick_scan': url_quick_scan,
+    'quick_scan_by_id': quick_scan_by_id,
     'submit_url': submit_url
 
 }