+++ hybrid-analysis/release_notes.md
@@ -1,5 +1,7 @@
-#### Following enhancements have been made to the Hybrid Analysis Connector in version 2.0.2:
+#### Following enhancements have been made to the Hybrid Analysis Connector in version 2.1.0:
 
+- Added the following action and it's playbook:
+    - Quick Scan by ID
 - Renamed the action and playbook label from `Get Analysis Report for Multiple Hashcodes` to
   `Get Analysis Report for Hashcode`.
 - Renamed the `Hash Codes` parameter to `Hash Code` for the action `Get Analysis Report for Hashcode`.